Complet game
* Move state from Board to Game
* Store history in Game state
* Game passes the currently selected array of squares to Board
* Board displays the squares in rows by passing each to the Square component
* Board also passes the onClick function (rec’vd from Game) to Square
