+++ ttt/src/index.js
@@ -3,95 +3,7 @@ import ReactDOM from 'react-dom';
 import Game from './Game';
 import './index.css';
 
-class Square extends React.Component {
-  render() {
-    return (
-      <button
-        className="square"
-        onClick={() => this.props.onClick()}
-      >
-        {this.props.value}
-      </button>
-    );
-  }
-}
 
-class Board extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      squares: Array(9).fill(null),
-    }
-  }
-
-  handleClick(i) {
-    const squares = this.state.squares.slice()
-    const turnCount = squares.filter(sq => sq === null).length
-
-    console.log('turnCount: ', turnCount);
-    // console.log('squares: ', squares);
-    console.log('state before: ', this.state.squares);
-
-    turnCount % 2 !== 0 && squares[i] === null ? squares[i] = "X" : squares[i] = "O"
-    this.setState({ squares: squares })
-    // console.log('state after: ', this.state.squares);
-  }
-
-  componentDidUpdate() {
-    console.log('state after update: ', this.state.squares);
-  }
-
-
-  renderSquare(i) {
-    return <Square
-      value={this.state.squares[i]}
-      onClick={() => this.handleClick(i)}
-    />;
-  }
-
-  render() {
-    const status = 'Next player: X';
-
-    return (
-      <div>
-        <div className="status">{status}</div>
-        <div className="board-row">
-          {this.renderSquare(0)}
-          {this.renderSquare(1)}
-          {this.renderSquare(2)}
-        </div>
-        <div className="board-row">
-          {this.renderSquare(3)}
-          {this.renderSquare(4)}
-          {this.renderSquare(5)}
-        </div>
-        <div className="board-row">
-          {this.renderSquare(6)}
-          {this.renderSquare(7)}
-          {this.renderSquare(8)}
-        </div>
-      </div>
-    );
-  }
-}
-
-class Game extends React.Component {
-  render() {
-    return (
-      <div className="game">
-        <div className="game-board">
-          <Board />
-        </div>
-        <div className="game-info">
-          <div>{/* status */}</div>
-          <ol>{/* TODO */}</ol>
-        </div>
-      </div>
-    );
-  }
-}
-
-// ========================================
 
 ReactDOM.render(
   <Game />,
